Move Metal API interfaces to RHI
