+++ taichi/rhi/metal/CMakeLists.txt
@@ -4,6 +4,7 @@ set(METAL_RHI metal_rhi)
 add_library(${METAL_RHI})
 target_sources(${METAL_RHI}
   PRIVATE
+    api.cpp
     device.cpp
   )
